add DelegatingTracker and NoOpTracker. Extract logic from LocalTracker to JUnitTracker and DeviceTracker
